autoscaler/main.py
"""
GitHub Actions Autoscaler for Cloud Run Worker Pools

This Cloud Run function handles GitHub webhook events (workflow_job)
and dynamically scales the worker pool based on demand.

TODO: Full implementation tracked in Issue #X
"""

import os
import json
import hmac
import hashlib
from flask import Flask, request, jsonify
from google.cloud import run_v2

app = Flask(__name__)

# Configuration
PROJECT_ID = os.environ.get("PROJECT_ID")
REGION = os.environ.get("REGION", "us-central1")
WORKER_POOL_NAME = os.environ.get("WORKER_POOL_NAME", "github-runners")
WEBHOOK_SECRET = os.environ.get("GITHUB_WEBHOOK_SECRET")
MAX_INSTANCES = int(os.environ.get("MAX_INSTANCES", "10"))


def verify_webhook_signature(payload: bytes, signature: str) -> bool:
    """Verify the GitHub webhook signature."""
    if not WEBHOOK_SECRET:
        return True  # Skip verification if no secret configured

    expected = hmac.new(
        WEBHOOK_SECRET.encode(),
        payload,
        hashlib.sha256
    ).hexdigest()

    return hmac.compare_digest(f"sha256={expected}", signature)


def scale_worker_pool(target_instances: int) -> dict:
    """Scale the Cloud Run worker pool to target instance count."""
    # TODO: Implement Cloud Run worker pool scaling
    # Uses google.cloud.run_v2.WorkerPoolsClient
    pass


@app.route("/webhook", methods=["POST"])
def handle_webhook():
    """Handle incoming GitHub webhook events."""
    # Verify signature
    signature = request.headers.get("X-Hub-Signature-256", "")
    if not verify_webhook_signature(request.data, signature):
        return jsonify({"error": "Invalid signature"}), 401

    # Parse event
    event_type = request.headers.get("X-GitHub-Event")
    payload = request.json

    if event_type != "workflow_job":
        return jsonify({"status": "ignored", "reason": "not workflow_job"})

    action = payload.get("action")

    if action == "queued":
        # Job queued - scale up
        # TODO: Implement scale-up logic
        pass
    elif action in ("completed", "cancelled"):
        # Job finished - potentially scale down
        # TODO: Implement scale-down logic
        pass

    return jsonify({"status": "processed", "action": action})


@app.route("/health", methods=["GET"])
def health():
    """Health check endpoint."""
    return jsonify({"status": "healthy"})


if __name__ == "__main__":
    app.run(host="0.0.0.0", port=int(os.environ.get("PORT", 8080)))
