autoscaler/requirements.txt
flask>=2.3.0
google-cloud-run>=0.10.0
gunicorn>=21.0.0
